style(dashboard): Align 'add-a-review' button and searchbar
